refactor: context to avoid errors
