+++ src/components/ThemeSelector/ThemeSelector.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useTheme } from '../../contexts/ThemeContext';
+import useTheme from '../../hooks/useTheme';
 
 const ThemeSelector: React.FC = () => {
   const { theme, setTheme } = useTheme();
